refactor(platform-bitbucket-server): 适配器按领域拆分为四个服务
重构 Phase 2c（Bitbucket）。把单体 BitbucketServerAdapter 拆为四个领域服务
（BitbucketServerConnection / BitbucketPullRequestService / BitbucketCommentService /
BitbucketMediaService，各 extends platform-core 领域基类），共享一份连接上下文；
BitbucketServerAdapter 退化为薄门面，过渡期委托扁平方法并暴露 .connection/.pulls/
.comments/.media。

- 连接配置归连接层自管：client 自管 clone 协议、webBase 与 getCloneUrl；各服务只持 client。
- 映射就近维护：领域专属映射下沉为各服务私有方法（版本比较、合并状态、评论 / 锚点映射等），
  公共响应类型抽到 types.ts，跨领域 mapUser 放 utils.ts。
- review 决断态 / 活动 action 映射改常量表，方法补多行 JSDoc。

公开方法签名不变、适配器测试全过、desktop typecheck 通过，行为等价。
